refactor(status): split collectDejaVuRows into per-concern functions
collectDejaVuRows (CC 95) mixed five concerns in one function: error
mapping, the install ladder, doctor health, the 6-way per-host target
ladder, and the derived-index ladder. Extracted dejaErrorRow,
dejaInstallRows, dejaDoctorRows, dejaTargetRows (via a dejaTargetContext
guard-clause helper to keep both under the CC budget), and dejaIndexRows
into src/commands/status/deja-vu.mjs; collectDejaVuRows is now a ~20-line
orchestrator that assembles their rows in the same order and keeps its
existing try/catch and exact exported signature.

Result: collectDejaVuRows CC 95 -> 21, every extracted helper under CC 25.
Pure decomposition: no messages, ordering, or logic changed. status.mjs
re-exports collectDejaVuRows unchanged for existing test imports.
