+++ src/commands/status.mjs
@@ -32,8 +32,9 @@ import { removedAgentGaps, upstreamFixAvailable } from '../lib/scaffold.mjs';
 import { latestSnapshot, readModelStore, summarizeModelHealth } from '../lib/model-inventory/index.mjs';
 import { row } from './status/row.mjs';
 import { renderHostDetailRows, admittedLifecycleFallbackRows } from './status/host-detail.mjs';
+import { collectDejaVuRows } from './status/deja-vu.mjs';
 
-export { renderHostDetailRows };
+export { renderHostDetailRows, collectDejaVuRows };
 
 export const options = {
   json: { type: 'boolean', default: false },
@@ -57,166 +58,6 @@ Examples:
   ak status           quick dashboard
   ak status --deep    thorough check
   ak status --json    machine-readable rows`;
-
-const DEJA_HOSTS = Object.freeze(['claude', 'codex', 'opencode']);
-const SAFE_VERSION = /^v?\d+\.\d+\.\d+(?:-[0-9A-Za-z.-]+)?(?:\+[0-9A-Za-z.-]+)?$/;
-
-function hasDejaVuOwnership(cfg) {
-  const ownership = cfg?.integrations?.ownership?.dejaVu;
-  return !!ownership?.install
-    || (!!ownership?.targets && typeof ownership.targets === 'object'
-      && Object.keys(ownership.targets).length > 0);
-}
-
-function safeDejaCode(value) {
-  return typeof value === 'string' && /^[a-z0-9-]{1,80}$/.test(value)
-    ? value : 'unavailable';
-}
-
-function safeDejaVersion(value) {
-  return typeof value === 'string' && SAFE_VERSION.test(value) ? value : 'unknown';
-}
-
-/**
- * Render the managed companion from its bounded lifecycle facts. No upstream
- * path, command output, signature, transcript metadata, or plugin payload is
- * copied into a row. A fix is attached only when the same adapter plan contains
- * an operation that can perform it.
- * @param {{cfg?:any,adapter?:any,planOptions?:Record<string,any>}} [options]
- */
-export async function collectDejaVuRows(options = {}) {
-  const {
-    cfg,
-    adapter = companionLifecycleFor('deja-vu'),
-    planOptions = {},
-  } = options;
-  const desired = cfg?.integrations?.tools?.dejaVu;
-  const enabled = desired?.enabled === true;
-  const owned = hasDejaVuOwnership(cfg);
-  if (!enabled && !owned) {
-    return [row('deja-vu', 'info',
-      'deja-vu disabled — package, host wiring, and history remain unprobed')];
-  }
-  if (!adapter) {
-    return [row('deja-vu', 'warn', 'deja-vu lifecycle adapter unavailable')];
-  }
-
-  try {
-    const facts = await adapter.detect({ cfg });
-    const plan = await adapter.plan({ cfg, facts, options: planOptions });
-    const operations = Array.isArray(plan?.operations) ? plan.operations : [];
-    const actionable = !facts?.error && !plan?.error;
-    const hasOperation = (kind, host = null) => actionable && operations.some((operation) =>
-      operation?.kind === kind && (host === null || operation.host === host));
-    const rows = [];
-
-    if (facts?.error || plan?.error) {
-      const code = safeDejaCode(facts?.error ?? plan?.error);
-      const external = code === 'deja-external-version-unsupported'
-        || code === 'deja-external-install-unusable';
-      rows.push(row('deja-vu', external ? 'warn' : 'fail', external
-        ? `external deja-vu installation is not safely manageable (${code}); preserved`
-        : `deja-vu health contract failed closed (${code})`));
-    }
-
-    const install = facts?.install ?? {};
-    if (install.binaryPresent === false) {
-      if (enabled && hasOperation('package-install')) {
-        rows.push(row('deja-vu', 'warn', 'deja-vu package missing',
-          'sync installs the managed npm companion'));
-      } else if (install.ownership === 'external') {
-        rows.push(row('deja-vu', 'warn', 'external deja-vu package is unavailable; preserved'));
-      } else {
-        rows.push(row('deja-vu', 'info', 'deja-vu package absent'));
-      }
-    } else if (install.binaryPresent === true) {
-      const version = safeDejaVersion(install.version);
-      if (hasOperation('package-upgrade')) {
-        rows.push(row('deja-vu', 'warn',
-          `managed deja-vu ${version} has an available package upgrade`,
-          'sync upgrades the owned npm companion'));
-      } else if (install.receiptState === 'drifted' || install.receiptState === 'malformed') {
-        rows.push(row('deja-vu', 'warn',
-          `deja-vu ${version} package ownership receipt drifted; current installation preserved`));
-      } else if (install.ownership === 'agentic-kit') {
-        rows.push(row('deja-vu', install.supported === false ? 'warn' : 'ok',
-          `managed deja-vu ${version} package${install.supported === false ? ' is below v0.19.0' : ' is present'}`));
-      } else {
-        rows.push(row('deja-vu', 'info',
-          `external deja-vu ${version} package detected; installation remains user-owned`));
-      }
-    }
-
-    if (facts?.doctor?.state === 'ok') {
-      if (facts.doctor.health?.state === 'degraded') {
-        const issueCount = Number.isSafeInteger(facts.doctor.health.storeIssues)
-          ? Math.min(facts.doctor.health.storeIssues, 999) : 0;
-        rows.push(row('deja-vu', 'warn',
-          `deja-vu doctor schema v2 accepted but component health is degraded${issueCount > 0
-            ? ` (${issueCount} bounded store issue${issueCount === 1 ? '' : 's'})` : ''}`));
-      } else {
-        rows.push(row('deja-vu', 'ok', 'deja-vu doctor schema v2 accepted'));
-      }
-    }
-
-    for (const host of DEJA_HOSTS) {
-      const target = facts?.targets?.[host];
-      const receipt = cfg?.integrations?.ownership?.dejaVu?.targets?.[host];
-      const receiptPresent = !!receipt;
-      if (!target || (!target.selected && !receiptPresent)) continue;
-      const hasTargetOperation = hasOperation('target-remove', host)
-        || hasOperation('target-install', host);
-      const managedTransition = hasTargetOperation && receipt?.mode
-        && receipt.mode !== facts?.desired?.mode;
-      if (target.receiptState === 'drifted') {
-        rows.push(row('deja-vu', 'warn',
-          `${host}: managed target ownership drifted; current wiring preserved`));
-      } else if (target.conflict === 'external-auto-active' && !managedTransition) {
-        rows.push(row('deja-vu', 'warn',
-          `${host}: external automatic recall conflicts with MCP-only intent; external state preserved`));
-      } else if (target.satisfied) {
-        rows.push(row('deja-vu', target.ownership === 'agentic-kit' ? 'ok' : 'info',
-          `${host}: ${target.desiredTarget ?? 'deja-vu'} target active`
-          + (target.ownership === 'agentic-kit' ? ' and receipt-owned' : ' via external wiring')));
-      } else if (hasTargetOperation) {
-        rows.push(row('deja-vu', 'warn',
-          `${host}: managed deja-vu target requires convergence`,
-          `sync converges the exact ${host} target`));
-      } else if (target.hostPresent === false) {
-        rows.push(row('deja-vu', 'warn',
-          `${host}: selected host is unavailable; companion wiring skipped`));
-      } else {
-        rows.push(row('deja-vu', 'warn',
-          `${host}: desired deja-vu target is not proven; external state preserved`));
-      }
-    }
-
-    if (enabled && facts?.desired?.indexOnSetup) {
-      if (facts?.index?.state === 'ok') {
-        rows.push(row('deja-vu', 'ok', 'deja-vu derived index is healthy'));
-      } else if (hasOperation('index')) {
-        rows.push(row('deja-vu', 'warn',
-          `deja-vu derived index is ${['missing', 'stale'].includes(facts?.index?.state)
-            ? facts.index.state : 'not ready'}`,
-          'sync runs one bounded deja index after target convergence'));
-      } else if (facts?.index?.state === 'stale-readonly') {
-        rows.push(row('deja-vu', 'warn',
-          'deja-vu derived index is stale-readonly; automatic repair is unsafe'));
-      } else if (facts?.index?.state !== undefined) {
-        rows.push(row('deja-vu', 'info', 'deja-vu derived index health is unknown'));
-      }
-    }
-
-    if (plan?.warnings?.includes('deja-package-latest-unavailable')) {
-      rows.push(row('deja-vu', 'warn',
-        'managed deja-vu package is usable, but npm latest-version drift could not be verified'));
-    }
-
-    return rows.length ? rows : [row('deja-vu', 'info', 'deja-vu state is unobserved')];
-  } catch {
-    return [row('deja-vu', 'warn', 'deja-vu status unavailable')];
-  }
-}
 
 export async function collect({
   pkgRoot,
